+++ metadata.rb
@@ -8,6 +8,4 @@ version          '0.1.0'
 
 depends 'nfs', '~> 2.1.0'
 depends 'apt', '~> 2.9.2'
-depends 'build-essential', '~> 2.2.4'
-depends 'awscli', '~> 1.1.2'
 depends 'cron', '~> 1.6.1'