+++ recipes/create-metrics-dependencies.rb
@@ -1,7 +1,7 @@
 # Cookbook Name:: mh-opsworks-recipes
 # Recipe:: create-metrics-dependencies
 
-include_recipe "awscli::default"
+include_recipe "mh-opsworks-recipes::install-awscli"
 
 user "custom_metrics" do
   comment 'The custom metrics reporting user'