+++ recipes/populate-maven-cache.rb
@@ -1,7 +1,7 @@
 # Cookbook Name:: mh-opsworks-recipes
 # Recipe:: populate-maven-cache
 
-include_recipe "awscli::default"
+include_recipe "mh-opsworks-recipes::install-awscli"
 
 bucket_name = node.fetch(:shared_asset_bucket_name, 'mh-opsworks-shared-assets')